refactor: split read.go into taskfile.go and reader.go
